chore: rename register page
