+++ src/pages/register/index.ts
